% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_class.R
\name{point_class}
\alias{point_class}
\title{Title}
\usage{
point_class(
  point_data,
  higher_geo_lay,
  unique_id_code,
  class_col,
  crs,
  total_points = TRUE
)
}
\arguments{
\item{total_points}{}
}
\value{

}
\description{
Title
}
