% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points}
\alias{points}
\title{Point geospatial layer.}
\format{
A geospatial file of ten points georeferenced in wgs84.
}
\source{
Own dataset.
}
\usage{
points
}
\description{
Toy dataset of point data.
}
\keyword{datasets}
