% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pol_large}
\alias{pol_large}
\title{Large polygons geospatial layer.}
\format{
A geospatial file of three polygons georeferenced in wgs84.
}
\source{
Own dataset.
}
\usage{
pol_large
}
\description{
Toy dataset of polygon data.
}
\keyword{datasets}
