% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pol_small}
\alias{pol_small}
\title{Small polygons geospatial layer.}
\format{
A geospatial file of eight polygons georeferenced in wgs84.
}
\source{
Own dataset.
}
\usage{
pol_small
}
\description{
Toy dataset of polygon data.
}
\keyword{datasets}
