% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal_calc.R
\name{areal_calc}
\alias{areal_calc}
\title{Areal data calculation}
\usage{
areal_calc(polygon_layer, higher_geo_lay, unique_id_code, crs)
}
\arguments{
\item{polygon_layer}{multipologon object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{higher_geo_lay}{multipologon object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{unique_id_code}{a string, indicating the unique ID column of \code{higher_geo_lay} in which
we want to summarise the data}

\item{crs}{coordinate reference system: integer with the EPSG code, or character with proj4string}
}
\value{
a \code{tibble} data frame object
}
\description{
Calculates the area covered by
}
\examples{


}
