% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal_calc.R
\name{areal_calc}
\alias{areal_calc}
\title{Areal data calculation}
\usage{
areal_calc(polygon_layer, higher_geo_lay, unique_id_code, crs)
}
\arguments{
\item{crs}{}
}
\value{

}
\description{
Calculates the area covered by
}
