% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lines}
\alias{lines}
\title{Line geospatial layer.}
\format{
A geospatial file of six lines georeferenced in wgs84.
}
\source{
Own dataset.
}
\usage{
lines
}
\description{
Toy dataset of line data.
}
\keyword{datasets}
